+++ tests/workflows/test_forcing.py
@@ -5,15 +5,13 @@ import pytest
 import xarray as xr
 
 import hydromt._compat as compat
-from hydromt.data_catalog import DataCatalog
 from hydromt.gis.raster import full_from_transform
 from hydromt.workflows.forcing import pet, precip
 
 
 @pytest.mark.skip(reason="Needs implementation of all raster Drivers.")
-def test_precip():
-    cat = DataCatalog()
-    p_precip = cat.get_rasterdataset("era5", variables="precip")  # era hourly
+def test_precip(data_catalog):
+    p_precip = data_catalog.get_rasterdataset("era5", variables="precip")  # era hourly
 
     # We create a more refined grid
     p_transform = p_precip.raster.transform
@@ -27,7 +25,7 @@ def test_precip():
     xr.testing.assert_equal(pout.x, grid.x)
 
     # Testing with clim argument
-    p_clim = cat.get_rasterdataset("worldclim")
+    p_clim = data_catalog.get_rasterdataset("worldclim")
     # give it a nodata value in the datacatalog >> issue to
     # create for the data artifacts
     p_clim.raster.set_nodata(-999.0)
@@ -43,10 +41,9 @@ def test_precip():
 
 @pytest.mark.skip(reason="Needs implementation of all raster Drivers.")
 @pytest.mark.skipif(not compat.HAS_PYET, reason="pyET not installed.")
-def test_pet():
-    cat = DataCatalog()
-    et_data = cat.get_rasterdataset("era5_daily_zarr")
-    dem = cat.get_rasterdataset("era5_orography").squeeze("time", drop=True)
+def test_pet(data_catalog):
+    et_data = data_catalog.get_rasterdataset("era5_daily_zarr")
+    dem = data_catalog.get_rasterdataset("era5_orography").squeeze("time", drop=True)
 
     peto = pet(et_data, et_data, dem, method="penman-monteith_tdew")